+++ R/mapvalues.R
@@ -1,7 +1,0 @@
-#' @title Replace specified values in a vector or factor with new values.
-#' @description Same function as in \pkg{plyr}. Imported by \pkg{FSA} and exported to minimize conflicts with other commonly used packages. See \code{\link[plyr]{mapvalues}} for documentation.
-#' @name mapvalues
-#' @rdname mapvalues
-#' @importFrom plyr mapvalues
-#' @export
-NULL